requirements.txt
hydralit_components==1.0.10
mediapipe==0.10.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
Requests==2.31.0
SpeechRecognition==3.10.0
streamlit==1.23.1
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
streamlit_toggle_switch==1.0.2
